+++ packages/solvejs/src/index.ts
@@ -6,3 +6,4 @@ export * from "@jdsalasc/solvejs-constants";
 export * from "@jdsalasc/solvejs-numbers";
 export * from "@jdsalasc/solvejs-validators";
 export * from "@jdsalasc/solvejs-objects";
+export * from "@jdsalasc/solvejs-async";